requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
PyYAML==6.0.3
pytest==9.0.2
mypy==1.18.2
ruff==0.13.2
